Stream is now a specialized wrapper of Continuation
Turns out, the old way of doing streams was identical to continuations.
The source of the continuation just needs to call the handler with each
element, and the return type has to be isomorphic to the unit type (),
so that something can always be returned. That isomorphic type is
Future<()>, to support streams being asynchronous.